Bugfix and improvement for start of panning
Fixed obvious mistake of subtracting touchCenterCoord_ from new
start coordinate when starting a pan. This lead to a jump to a location
close to 0,0 when touchCenterCoord_ was set.
Since lastPos_ is also maintained as the center between two touch
points, smooth start could also be done in that case.

Change-Id: Icf8776aeb1ea3ae0cbadc74a28e2db22058f05d1
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>

Conflicts:
	src/imports/location/qdeclarativegeomapgesturearea.cpp