Fix issues with MapItemView.
MapItemView did not allow clearing either the model or the delegate.
There is no reason to restrict clearing these properties. Doing so
makes it difficult to temporarily hide the view setting a null model.

Signal connections between the model and the view were never
disconnected, which could result in strange behavior if a single view
is used to display multiple models.

Removed unneeded member variables and declared but unimplemented member
functions.

Change-Id: Ife4745ab38104460960c18419ed1b95e72ffab23
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>