[mer] Limit the minimum zoom level based on tile size and map item size. Contributes to MER#1122
Depending on the size of the map item the minimum zoom supported by the
plugin may be such that empty gray borders are shown around the map when
zoomed out. Limit the minimum zoom, for tiled maps, so that those
borders are not shown.

A related problem also exists which is not affected by this change. It
is not always possible to zoom out far enough to see the full view of
at least one dimension. Fixing this is a more extensive change requiring
supporting zoom levels less than supported by the plugin, even negative
zoom levels.

[ChangeLog][QtLocation][Maps] The minimum zoom level for tile based
maps is now limited based on the size of the map item and the tile size.

Change-Id: I667e86c15a30068198a034aa5020d6a6b99f46b6