[mer] Fix issue rendering tiles from incorrect zoom level. Contributes to MER#1122
Prefetching tiles fetches tiles from adjacent zoom levels and updates
the state of QGeoCameraTiles, but does not change the camera. If the
next rendering pass does not update the camera the set of visible tiles
will not be updated, causing the render to use the prefetch tiles
instead of the visible tiles.

Fixed by making QGeoCameraTiles::prefetchTiles() const so that it does
not invalidate the set of visible tiles.

Make QGeoCameraTiles aware of the minimum zoom level to prevent
requesting invalid tiles from the tile provider.

Change-Id: Ifcde130eea5d4f9f0bb5449b804943eadd75512b